requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    cd ~/src/kubecost && run-piptools
#
ruamel-yaml==0.17.40
    # via -r requirements.in
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml

pip==23.0.1
    # via -r piptools_requirements.txt
setuptools==67.6.0
    # via -r piptools_requirements.txt